Changed two instances of unnecessarily convoluted hash initializations
I could see no reason for these to be written the way they were. The block passed to Hash.new would never be run unless the constant was modified later, and I see no reason why it should. There's no test that modifies them, at least. I froze the hashes to make sure that if it's the case that they are modified, this will become obvious.