Fix Ruby 3.5 compatibility
In Ruby 3.5 most of the cgi gem will be removed. Only the various escape/unescape methods will be retained by default.

For `CGI.parse`, there is a mostly equivalent method from `URI.decode_www_form`. The data is in a slightly different
format, so it needs to be converted.
For `CGI.unescape`, depending on the Ruby version. different files need to be required.

https://bugs.ruby-lang.org/issues/21258